parser: Move parsers out of extras sub-package
The parsers are important enough to warrant their own sub-package.
